refactor: decompose BlobStore into single-concern modules (v1.4.0)
