+++ test/blob-expiration.test.ts
@@ -4,7 +4,7 @@ import { tmpdir } from "node:os";
 import { join } from "node:path";
 import { afterEach, describe, expect, it } from "vitest";
 import { BlobHistoryStore, blobTombstonePath } from "../src/core/blob-history-store.js";
-import { BlobStore } from "../src/core/blob-store.js";
+import { BlobStore } from "../src/core/blob-store/index.js";
 
 const temporaryDirectories: string[] = [];
